Started refactoring view
