+++ core/src/main/java/com/g16/feyrune/controller/CombatInputProcessor.java
@@ -3,7 +3,6 @@ package com.g16.feyrune.controller;
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Input;
 import com.g16.feyrune.enums.Direction;
-import com.g16.feyrune.view.combat.CombatInputHandler;
 
 public class CombatInputProcessor implements IInput {
     private CombatInputHandler inputHandler;
