+++ core/src/main/java/com/g16/feyrune/controller/combat/ChoiceRenderer.java
@@ -1,13 +1,14 @@
-package com.g16.feyrune.view.combat;
+package com.g16.feyrune.controller.combat;
 
+import com.g16.feyrune.controller.combat.ChoiceButton;
 import com.g16.feyrune.enums.Direction;
 import com.g16.feyrune.enums.Selection;
 
-import java.util.ArrayList;
+import java.util.List;
 
 public class ChoiceRenderer {
     private Selection currentSelection;
-    private ArrayList<ChoiceButton> buttons;
+    private List<ChoiceButton> buttons;
 
     public ChoiceRenderer(){ //TODO: there should be some logic here to select one of them
         buttons.add(new ChoiceButton());
