--- core/src/main/java/com/g16/feyrune/interfaces/IScene.java
@@ -1,7 +1,0 @@
-package com.g16.feyrune.interfaces;
-
-public interface IScene {
-    void update();
-
-    void render();
-}
