+++ core/src/main/java/com/g16/feyrune/view/IScene.java
@@ -1,0 +1,9 @@
+package com.g16.feyrune.view;
+
+import com.badlogic.gdx.graphics.g2d.SpriteBatch;
+
+public interface IScene {
+    void update();
+
+    void render(SpriteBatch batch);
+}
