+++ core/src/main/java/com/g16/feyrune/view/View.java
@@ -1,11 +1,8 @@
 package com.g16.feyrune.view;
 
-import com.badlogic.gdx.graphics.Camera;
-import com.badlogic.gdx.graphics.OrthographicCamera;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.g16.feyrune.model.ModelState;
 import com.g16.feyrune.interfaces.IObserver;
-import com.g16.feyrune.interfaces.IScene;
 import com.g16.feyrune.model.Model;
 import com.g16.feyrune.view.combat.CombatScene;
 import com.g16.feyrune.view.overworld.OverworldScene;
@@ -22,8 +19,8 @@ public class View implements IObserver {
         this.model.registerNewObserver(this);
 
         batch = new SpriteBatch();
-        overworldScene = new OverworldScene(model.getPlayer(),batch);
-        combatScene = new CombatScene(batch);
+        overworldScene = new OverworldScene(model.getPlayer());
+        combatScene = new CombatScene();
 
         update(); //TODO: this is also wrong and bad :)
     }
